+++ code/boss-api/trunk/BOSS-API.h
@@ -188,7 +188,7 @@ BOSS_API uint32_t TrialSortMods(boss_db db, uint8_t ** sortedPlugins, const uint
 // Returns an array of the Bash Tags encounterred when loading the masterlist
 // and userlist, and the number of tags in the returned array. The array and
 // its contents are static and should not be freed by the client.
-BOSS_API uint32_t GetBashTagMap (boss_db db, BashTag ** tagMap, uint32_t * numTags);
+BOSS_API uint32_t GetBashTagMap (boss_db db, BashTag ** tagMap, size_t * numTags);
 
 // Returns arrays of Bash Tag UIDs for Bash Tags suggested for addition and removal 
 // by BOSS's masterlist and userlist, and the number of tags in each array.
@@ -199,9 +199,9 @@ BOSS_API uint32_t GetBashTagMap (boss_db db, BashTag ** tagMap, uint32_t * numTa
 // suggestion message additions.
 BOSS_API uint32_t GetModBashTags (boss_db db, const uint8_t * modName, 
 									uint32_t ** tagIds_added, 
-									uint32_t * numTags_added, 
+									size_t * numTags_added, 
 									uint32_t **tagIds_removed, 
-									uint32_t *numTags_removed,
+									size_t *numTags_removed,
 									bool * userlistModified);
 									
 // Returns the message associated with a dirty mod and whether the mod needs
